src/requirements.txt
anyjson==0.3.3
Django==1.6
django-braces==1.0.0
django-nose==1.2
djangorestframework==2.3.10
httplib2==0.8
Mako==0.9.0
MarkupSafe==0.19
pytz
requests==2.2.1
simplejson==3.3.3
six==1.5.2
unittest2==0.5.1
urllib3==1.9
astroid==1.3.4
coverage==3.7.1
django-jenkins==0.16.4
flake8==2.3.0
logilab-astng==0.24.3
logilab-common==0.63.2
mccabe==0.3
pep8==1.5.7
pyflakes==0.8.1
pylint==1.4.1
unittest-xml-reporting==1.10.0
django-extensions==1.5.0
graphviz==0.4.2
epydoc==3.0.1
